+++ MapObjectProperties/MassProperty.cs
@@ -1,5 +1,4 @@
 ﻿using MapsExt.Properties;
-using System;
 using UnboundLib;
 using UnityEngine;