+++ VanillaMapObjectsEditor/MapObjectProperties/EditorBoxHealthProperty.cs
@@ -1,19 +1,10 @@
 ﻿using MapsExt;
 using MapsExt.Editor.Properties;
 using MapsExt.Editor.UI;
-using MapsExt.Editor.Utils;
 using MapsExt.Properties;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using UnboundLib;
 using UnityEngine;
-using UnityEngine.SocialPlatforms;
 using VanillaMapObjects.MapObjectProperties;
-using VanillaMapObjects.MapObjects;
-using VanillaMapObjectsEditor.MapObjects;
 
 namespace VanillaMapObjectsEditor.MapObjectProperties
 {