+++ VanillaMapObjectsEditor/Visualizers/IMapObjectVisualizer.cs
@@ -1,0 +1,7 @@
+﻿namespace MapsExt.Visualizers
+{
+    interface IMapObjectVisualizer
+    {
+        void SetEnabled(bool enabled);
+    }
+}