+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+#### 1.1.0:
+- Added Springy Ropes
+
 #### 1.0.0:
 - Released