+++ VanillaMapObjectsEditor/CHANGELOG.md
@@ -1,3 +1,6 @@
+#### 1.1.0:
+- Added Springy Ropes
+
 #### 1.0.1:
 - Uploaded the correct DLL